# pnmc-api/openapi.yaml
openapi: 3.0.1
info:
  title: PNMC.Api
  version: '1.0'
paths:
...
  /api/v1/institucional/categorias-contenido:
    get:
      tags:
        - categorias
      operationId: ListarCategorias
      parameters:
        - name: modulo
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      tags:
        - categorias
      operationId: CrearCategoria
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardarCategoriaSolicitud'
        required: true
      responses:
        '200':
          description: OK
  '/api/v1/institucional/categorias-contenido/{id}':
    put:
      tags:
        - categorias
      operationId: GuardarCategoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardarCategoriaSolicitud'
        required: true
      responses:
        '200':
          description: OK
    delete:
      tags:
        - categorias
      operationId: EliminarCategoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
  '/api/v1/institucional/categorias-contenido/{id}/fusionar':
    post:
      tags:
        - categorias
      operationId: FusionarCategoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FusionarCategoriaSolicitud'
        required: true
      responses:
        '200':
          description: OK
  /api/v1/publico/categorias-contenido:
    get:
      tags:
        - categorias-publico
      operationId: ListarCategoriasPublicas
      parameters:
        - name: modulo
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    FusionarCategoriaSolicitud:
      type: object
      properties:
        destinoId:
          type: integer
          format: int32
      additionalProperties: false
...
    GuardarCategoriaSolicitud:
      type: object
      properties:
        codigoModulo:
          type: string
          nullable: true
        nombreCategoria:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        descripcion:
          type: string
          nullable: true
        ordenVisualizacion:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
...